refactor: :recycle: more refactors to rust backend, seperate folders
