+++ shredr-backend/src/db/mod.rs
@@ -1,0 +1,5 @@
+pub mod db;
+pub mod db_routes;
+
+pub use db::{CreateBlobRequest, DbHandler};
+pub use db_routes::AppState;
