+++ shredr-backend/src/websocket/mod.rs
@@ -1,0 +1,4 @@
+pub mod websocket;
+pub mod websocket_routes;
+
+pub use websocket::{WebSocketMessage, WebSocketState};
